Update for Ruby >= 3 compatibility
Proc.new without a block used to capture a block passed to a method. This was removed in Ruby 3.

Previously it was a faster way to wrap a block and call it than the &block argument. `yield` is still the quickest way to invoke a passed block and `fetch` requires a block so we can safely call yield w/o checking